+++ src/main/java/nextstep/courses/domain/session/SessionCoverImage.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.exception.InvalidUrlPatternException;
 import nextstep.courses.utility.UrlPattern;
