+++ src/main/java/nextstep/courses/domain/session/SessionRegistrationBuilder.java
@@ -1,5 +1,6 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
+import nextstep.courses.domain.student.Students;
 import nextstep.users.domain.NsUser;
 
 public class SessionRegistrationBuilder {
