+++ src/main/java/nextstep/courses/infrastructure/JdbcStudentRepository.java
@@ -1,8 +1,8 @@
 package nextstep.courses.infrastructure;
 
-import nextstep.courses.domain.SessionRepository;
-import nextstep.courses.domain.Student;
-import nextstep.courses.domain.StudentRepository;
+import nextstep.courses.domain.session.SessionRepository;
+import nextstep.courses.domain.student.Student;
+import nextstep.courses.domain.student.StudentRepository;
 import nextstep.users.domain.UserRepository;
 import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.jdbc.core.RowMapper;
