+++ src/test/java/nextstep/courses/domain/session/SessionDurationTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.exception.IncorrectDurationException;
 import org.junit.jupiter.api.BeforeEach;
